added eslint with prettier
