+++ backend/src/createServer.js
@@ -8,12 +8,12 @@ function createServer() {
     typeDefs: 'src/schema.graphql',
     resolvers: {
       Mutation,
-      Query
+      Query,
     },
     resolverValidationOptions: {
-      requireResolversForResolveType: false
+      requireResolversForResolveType: false,
     },
-    context: req => ({ ...req, db })
+    context: req => ({ ...req, db }),
   });
 }
