+++ backend/src/db.js
@@ -4,7 +4,7 @@ const db = new Prisma({
   typeDefs: 'src/generated/prisma.graphql',
   endpoint: process.env.PRISMA_ENDPOINT,
   secret: process.env.PRISMA_SECRET,
-  debug: false
+  debug: false,
 });
 
 module.exports = db;
