+++ backend/src/index.js
@@ -1,6 +1,6 @@
 require('dotenv').config({ path: 'variables.env' });
 const createServer = require('./createServer');
-const db = require('./db');
+// const db = require('./db');
 
 const server = createServer();
 
@@ -8,8 +8,8 @@ server.start(
   {
     cors: {
       credentials: true,
-      origin: process.env.FRONTEND_URL
-    }
+      origin: process.env.FRONTEND_URL,
+    },
   },
   deets => {
     console.log(`Server running on http://localhost:${deets.port}`);
