+++ backend/src/util.js
@@ -1,0 +1,16 @@
+function hasPermission(user, permissionsNeeded) {
+  const matchedPermissions = user.permissions.filter(permissionTheyHave =>
+    permissionsNeeded.includes(permissionTheyHave)
+  );
+  if (!matchedPermissions.length) {
+    throw new Error(`You do not have sufficient permissions
+      : ${permissionsNeeded}
+
+      You Have:
+
+      ${user.permissions}
+      `);
+  }
+}
+
+exports.hasPermission = hasPermission;
