+++ frontend/components/Header.js
@@ -1,8 +1,8 @@
 import Link from 'next/link';
 import styled from 'styled-components';
-import Nav from './Nav';
 import Router from 'next/router';
 import NProgress from 'nprogress';
+import Nav from './Nav';
 
 Router.onRouteChangeStart = () => {
   NProgress.start();
